@@ -1,6 +1,6 @@
 {
   "name": "@impler/web",
-  "version": "0.27.1",
+  "version": "0.28.0",
   "author": "implerhq",
   "license": "MIT",
   "private": true,
@@ -27,6 +27,7 @@
     "@mantine/carousel": "6.0.21",
     "@mantine/core": "6.0.21",
     "@mantine/dates": "6.0.21",
+    "@mantine/dropzone": "6.0.21",
     "@mantine/hooks": "6.0.21",
     "@mantine/modals": "6.0.21",
     "@mantine/next": "6.0.21",
@@ -42,7 +43,7 @@
     "embla-carousel-react": "^7.0.9",
     "jwt-decode": "^3.1.2",
     "lottie-react": "^2.4.0",
-    "next": "14.1.1",
+    "next": "14.2.21",
     "react": "18.2.0",
     "react-ace": "^10.1.0",
     "react-chartjs-2": "^5.2.0",
