@@ -16,14 +16,13 @@ export class OnboardUser {
     private paymentAPIService: PaymentAPIService
   ) {}
 
-  async execute(command: OnboardUserCommand, email: string) {
+  async execute(command: OnboardUserCommand) {
     const createdProject = await this.createProject.execute(
       CreateProjectCommand.create({
         _userId: command._userId,
         name: command.projectName,
         onboarding: true,
-      }),
-      email
+      })
     );
 
     await this.userRepository.update(
