@@ -9,7 +9,8 @@ import { GetTemplates } from './get-templates/get-templates.usecase';
 import { GetEnvironment } from './get-environment/get-environment.usecase';
 
 import { UpdateImageColumns, SaveSampleFile } from '@shared/usecases';
-import { CreateTemplate, UpdateTemplateColumns, UpdateCustomization } from 'app/template/usecases';
+import { CreateTemplate, UpdateTemplateColumns, UpdateCustomization, AddColumn } from 'app/template/usecases';
+import { GetImportFileSchema } from 'app/template/usecases/get-column-types/get-import-file-schema.usecase';
 
 export const USE_CASES = [
   GetProjects,
@@ -24,8 +25,19 @@ export const USE_CASES = [
   UpdateCustomization,
   UpdateTemplateColumns,
   UpdateImageColumns,
+  GetImportFileSchema,
+  AddColumn,
   //
 ];
 
-export { CreateProject, GetProjects, UpdateProject, DeleteProject, GetTemplates, GetEnvironment, GetImports };
+export {
+  CreateProject,
+  GetProjects,
+  UpdateProject,
+  DeleteProject,
+  GetTemplates,
+  GetEnvironment,
+  GetImports,
+  AddColumn,
+};
 export { CreateProjectCommand, UpdateProjectCommand };
