@@ -1,6 +1,6 @@
 import { ApiService } from '@api';
 import { Dispatch, SetStateAction } from 'react';
-import { IUpload, WIDGET_TEXTS } from '@impler/client';
+import { IUpload, WIDGET_TEXTS, WidgetConfig } from '@impler/client';
 import { ITemplate, IImportConfig } from '@impler/shared';
 
 export interface IImplerStore {
@@ -29,6 +29,7 @@ export enum FlowsEnum {
 export interface IAppStore {
   title?: string;
   texts: typeof WIDGET_TEXTS;
+  config?: WidgetConfig;
   importId?: string;
   imageSchema?: string;
   data?: string;
